requirements.txt
python-vlc
windows-curses; sys_platform == 'win32'

requests~=2.32.3